brute force move
