+++ src/promptflow-tracing/promptflow/tracing/_version.py
@@ -2,6 +2,4 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
 
-import importlib.metadata
-
-__version__ = importlib.metadata.version("promptflow-tracing")
+VERSION = "0.0.1"
