+++ src/promptflow-tracing/tests/e2etests/simple_functions.py
@@ -63,10 +63,7 @@ def render_prompt_template(prompt: PromptTemplate, **kwargs):
 def openai_chat(connection: dict, prompt: str, stream: bool = False):
     client = AzureOpenAI(**connection)
 
-    messages = [
-        {"role": "system", "content": "You are a helpful assistant."},
-        {"role": "user", "content": prompt}
-    ]
+    messages = [{"role": "system", "content": "You are a helpful assistant."}, {"role": "user", "content": prompt}]
     response = client.chat.completions.create(model="gpt-35-turbo", messages=messages, stream=stream)
     return response.choices[0].message.content or ""
