--- src/promptflow/promptflow/tracing/_version.py
@@ -1,5 +1,0 @@
-# ---------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# ---------------------------------------------------------
-
-VERSION = "0.0.1"
